refactor(game): extract GamepadNavigationManager/NavLayer out of game_service
Move the two self-contained gamepad-focus classes (NavLayer +
GamepadNavigationManager) out of the game_service god-file into a new
lib/services/gamepad/gamepad_navigation_manager.dart. They are unrelated to
GameService's launch/session logic and reference nothing from it (only
LoggerService), so this is a pure (d) move — the first slice of the Phase 4
F game_service decomposition.

game_service.dart re-exports both symbols
(`export 'gamepad/gamepad_navigation_manager.dart' show
GamepadNavigationManager, NavLayer;`) and imports the file for its own
internal reactivate()/deactivateAll() calls, so all 32 importers (including
the 3 `show GamepadNavigationManager` importers) keep resolving them through
game_service.dart with zero changes. The library's public surface is
unchanged.

Bodies are byte-identical to the originals (verified vs main). Host −120.

Verified: dart format clean, flutter analyze No issues project-wide,
207/207 tests pass.
